#10: refactored code based on review.
