+++ app/src/main/res/layout/calendar_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragments.CalendarFragment">
+    tools:context=".Fragments.CalendarFragment.CalendarFragment">
 
     <TextView
         android:layout_width="match_parent"
